requirements.txt
numpy
scipy
matplotlib
sympy
whisper
soundfile
streamlit
pillow
